server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic==2.10.4
pydantic-settings==2.6.1
email-validator==2.2.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.2.3
google-generativeai==0.8.3
